+++ src/soho/field-filter/soho-field-filter.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SohoFieldFilterDirective } from './soho-field-filter.directive';
+
+@NgModule({
+  imports: [
+    CommonModule,
+  ],
+  declarations: [
+    SohoFieldFilterDirective
+  ],
+  exports: [
+    SohoFieldFilterDirective
+  ]
+})
+export class SohoFieldFilterModule {}